requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
azure-cognitiveservices-speech==1.30.0
beautifulsoup4==4.12.2
black==23.3.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.4
colorama==0.4.6
comtypes==1.2.0
dataclasses-json==0.5.9
docopt==0.6.2
faiss-cpu==1.7.4
Flask==2.3.2
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==20.1.0
html5lib==1.1
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.225
langchainplus-sdk==0.0.20
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.4
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pathspec==0.11.1
pipreqs==0.4.13
platformdirs==3.8.1
pydantic==1.10.11
python-dotenv==1.0.0
pyttsx3==2.90
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.18
tenacity==8.2.2
tiktoken==0.4.0
tomli==2.0.1
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.3
webencodings==0.5.1
Werkzeug==2.3.6
yarg==0.1.9
yarl==1.9.2
zipp==3.15.0
